update 機能ごとにファイル分割
ついでにlオプションのテスト対象のディレクトリを変更。理由はカレントディレクトリだと、lib以下を変更するとlオプションの結果が更新されてしまうため
