--- 08.ls_object/lib/print.rb
@@ -1,2 +1,0 @@
-#!/usr/bin/env ruby
-# frozen_string_literal: true
