+++ 08.ls_object/test/ls_test.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 require 'minitest/autorun'
-require_relative '../lib/ls'
+require_relative '../lib/file'
+require_relative '../lib/directory'
 
 class LsCommandTest < Minitest::Test
   def test_not_exist
@@ -60,15 +61,24 @@ class LsCommandTest < Minitest::Test
 
   def test_directory_l_option
     expected = <<~TEXT.chomp
-      total 16
-      -rw-r--r--  1 ryo  staff  141 11 25 23:38 Gemfile
-      -rw-r--r--  1 ryo  staff  878 11 25 23:38 Gemfile.lock
-      drwxr-xr-x  3 ryo  staff   96 11 25 22:37 bin
-      drwxr-xr-x  3 ryo  staff   96 11 25 21:52 doc
-      drwxr-xr-x  5 ryo  staff  160 11 26 16:24 lib
-      drwxr-xr-x  4 ryo  staff  128 11 25 23:38 test
+      total 8
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 1
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 1.rb
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 10
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 1111.rb
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 2
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 3
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 4
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 5
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 6
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 7
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 8
+      -rw-r--r--   1 ryo  staff   10 11 25 21:52 9
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 a.rb
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 bbbb.rb
+      drwxr-xr-x  13 ryo  staff  416 11 25 21:53 child_dir
     TEXT
-    assert_equal expected, ls_directories(%w[.], %w[-l], 13)
+    assert_equal expected, ls_directories(%w[test/test_data], %w[-l], 13)
   end
 
   def test_directory_a_option
@@ -82,39 +92,26 @@ class LsCommandTest < Minitest::Test
 
   def test_directory_rla_option
     expected = <<~TEXT.chomp
-      total 24
-      drwxr-xr-x   4 ryo  staff  128 11 25 23:38 test
-      drwxr-xr-x   5 ryo  staff  160 11 26 16:24 lib
-      drwxr-xr-x   3 ryo  staff   96 11 25 21:52 doc
-      drwxr-xr-x   3 ryo  staff   96 11 25 22:37 bin
-      -rw-r--r--   1 ryo  staff  878 11 25 23:38 Gemfile.lock
-      -rw-r--r--   1 ryo  staff  141 11 25 23:38 Gemfile
-      -rw-r--r--   1 ryo  staff   78 11 25 22:55 .rubocop.yml
-      drwxr-xr-x  15 ryo  staff  480 11 25 21:47 ..
-      drwxr-xr-x   9 ryo  staff  288 11 25 22:56 .
+      total 8
+      drwxr-xr-x  13 ryo  staff  416 11 25 21:53 child_dir
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 bbbb.rb
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 a.rb
+      -rw-r--r--   1 ryo  staff   10 11 25 21:52 9
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 8
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 7
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 6
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 5
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 4
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 3
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 2
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 1111.rb
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 10
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 1.rb
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 1
+      -rw-r--r--   1 ryo  staff    0 11 25 21:52 .test
+      drwxr-xr-x   4 ryo  staff  128 11 25 23:38 ..
+      drwxr-xr-x  18 ryo  staff  576 11 25 22:28 .
     TEXT
-    assert_equal expected, ls_directories(%w[.], %w[-l -a -r], 13)
+    assert_equal expected, ls_directories(%w[test/test_data], %w[-l -a -r], 13)
   end
-
-  # def test_ls
-  #   expected = <<~TEXT.chomp
-  #     Gemfile       bin           lib
-  #     Gemfile.lock  doc           test
-  #   TEXT
-  #   assert_equal expected, print_directories({:path=>".", :file_list=>["Gemfile", "Gemfile.lock", "bin", "doc", "lib", "test"]}, {:file=>[], :directory=>["."], :error=>[], :option=>[]})
-  # end
-
-  # def test_ls_input_path
-  #   expected = <<~TEXT
-  #   1 5 9
-  #   2 6 10
-  #   3 7
-  #   4 8
-  #   TEXT
-  #   assert_equal expected, main('sample_dir/hoge')
-  # end
-
-  # def test_ls_input_error_path
-  #   assert_equal 'ls: hoge: No such file or directory    ', main('hoge/')
-  # end
 end
